+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/meetingMember/dto/response/MeetingMemberResponseDto.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.meetingMember.dto.response;
+package com.mobble.mobbleserver.infrastructure.web.meetingMember.dto.response;
 
 import com.mobble.mobbleserver.domain.member.Member;
